% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malk.R
\docType{data}
\name{chalkboard}
\alias{chalkboard}
\title{a JSON file generated from a quarto chalkboard}
\format{
A character variable recognized by `jsonlite` as JSON.
}
\usage{
data(chalkboard)
}
\description{
The *chalkboard* data contains a character variable recognized by `jsonlite` as JSON.
}
\keyword{datasets}
