% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malk.R
\name{malk}
\alias{malk}
\title{Prepare data from a chalkboard.json file for plotting}
\usage{
malk(chalkboard_json_filepath, board_bg = "black", density = 10)
}
\arguments{
\item{chalkboard_json_filepath}{path to data that can be read with `jsonlite::fromJSON()`}

\item{board_bg}{string indicating the background color of the board}

\item{density}{integer indicating the number of points per pen stroke}
}
\value{
tibble with data formatted for `ggplot2` plotting
}
\description{
Formats json chalkboard data for plotting using the `ggplot2` package.
}
\details{
The output tibble creates points to corresponding to pen/chalk strokes, with the number of points indicated by the density.
}
\examples{

data(chalkboard)

malk(chalkboard)
}
\seealso{
\code{\link{plot_malk}}
}
