% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malk.R
\name{plot_malk}
\alias{plot_malk}
\title{Plot data from a chalkboard.json file that's been processed with the `malk` function}
\usage{
plot_malk(malk_output_tibble, board_bg = "black", point_size = 0.1)
}
\arguments{
\item{malk_output_tibble}{the output of the `malk` function, which has been applied to data that can be read with `jsonlite::fromJSON()`}

\item{board_bg}{string indicating the background color of the board}

\item{point_size}{float indicating the size of points in the penstroke}
}
\value{
theme added to a `ggplot2` object
}
\description{
Recreates a json chalkboard image using the `ggplot2` package.
}
\details{
The output `ggplot2` object uses points to recreate pen/chalk strokes, with the number of points indicated by the density.
}
\examples{
data(chalkboard)

malk(chalkboard) |> 
  plot_malk()
}
\seealso{
\code{\link{malk}}
}
